adding in put and patch calls for items

+++ WarhouseInventory.API/Models/Item.cs
@@ -25,7 +25,7 @@
 
         public string Supplier { get; set; }
 
-        public int SellingPrice { get; set; }
+        public double SellingPrice { get; set; }
 
     }
 }
