+++ WarhouseInventory.API/Models/ItemForCreation.cs
@@ -15,6 +15,6 @@ namespace WarehouseInventory.API.Models
 
         public string Supplier { get; set; }
 
-        public int SellingPrice { get; set; }
+        public double SellingPrice { get; set; }
     }
 }
